+++ engine/src/main/java/net/jqwik/engine/execution/lifecycle/RegisteredLifecycleHooks.java
@@ -1,10 +1,10 @@
 package net.jqwik.engine.execution.lifecycle;
 
 import net.jqwik.api.lifecycle.*;
-import net.jqwik.engine.LazyServiceLoaderCache;
+import net.jqwik.engine.support.*;
 
 public class RegisteredLifecycleHooks {
-	private static final LazyServiceLoaderCache<LifecycleHook> serviceCache = new LazyServiceLoaderCache(LifecycleHook.class);
+	private static final LazyServiceLoaderCache<LifecycleHook> serviceCache = new LazyServiceLoaderCache<>(LifecycleHook.class);
 
 	public static Iterable<LifecycleHook> getRegisteredHooks() {
 		return serviceCache.getServices();