+++ engine/src/main/java/net/jqwik/engine/providers/RegisteredArbitraryProviders.java
@@ -3,11 +3,11 @@ package net.jqwik.engine.providers;
 import java.util.*;
 
 import net.jqwik.api.providers.*;
-import net.jqwik.engine.LazyServiceLoaderCache;
+import net.jqwik.engine.support.*;
 
 public class RegisteredArbitraryProviders {
 
-	private static final LazyServiceLoaderCache<ArbitraryProvider> serviceCache = new LazyServiceLoaderCache(ArbitraryProvider.class);
+	private static final LazyServiceLoaderCache<ArbitraryProvider> serviceCache = new LazyServiceLoaderCache<>(ArbitraryProvider.class);
 
 	public static List<ArbitraryProvider> getProviders() {
 		return Collections.unmodifiableList(new ArrayList<>(serviceCache.getServices()));
@@ -23,14 +23,14 @@ public class RegisteredArbitraryProviders {
 	public static void unregister(ArbitraryProvider providerToDelete) {
 		List<ArbitraryProvider> services = serviceCache.getServices();
 		services.stream()
-			.filter(provider -> provider == providerToDelete)
-			.forEach(services::remove);
+				.filter(provider -> provider == providerToDelete)
+				.forEach(services::remove);
 	}
 
 	public static void unregister(Class<? extends ArbitraryProvider> providerClass) {
 		List<ArbitraryProvider> services = serviceCache.getServices();
 		services.stream()
-			.filter(provider -> provider.getClass() == providerClass)
-			.forEach(services::remove);
+				.filter(provider -> provider.getClass() == providerClass)
+				.forEach(services::remove);
 	}
 }