+++ engine/src/test/java/net/jqwik/api/providers/ProgrammaticArbitraryProviderRegistrationTests.java
@@ -6,8 +6,8 @@ import java.util.concurrent.*;
 import net.jqwik.api.*;
 import net.jqwik.engine.providers.*;
 
-import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.assertj.core.api.Assertions.assertThat;
+import static java.util.concurrent.TimeUnit.*;
+import static org.assertj.core.api.Assertions.*;
 
 @Group
 class ProgrammaticArbitraryProviderRegistrationTests implements AutoCloseable {
@@ -51,8 +51,8 @@ class ProgrammaticArbitraryProviderRegistrationTests implements AutoCloseable {
 		return !RegisteredArbitraryProviders.getProviders().contains(personProvider);
 	}
 
-	@Example
 	// only can fail if run as only/first test
+	@Example
 	void initIsSafe() throws Exception {
 		Callable<List<ArbitraryProvider>> lookup = RegisteredArbitraryProviders::getProviders;
 		ExecutorService exService = Executors.newFixedThreadPool(2);