+++ engine/src/main/java/net/jqwik/engine/execution/DefaultPropertyLifecycleContext.java
@@ -79,4 +79,8 @@ public class DefaultPropertyLifecycleContext extends AbstractLifecycleContext im
 		return parameterSupplierResolver.resolveParameter(executable, index, containerClass());
 	}
 
+	@Override
+	public PropertyAttributes attributes() {
+		return methodDescriptor.getConfiguration().getPropertyAttributes();
+	}
 }