+++ docs/release-notes.md
@@ -48,9 +48,9 @@ and
     - Shrunk samples have now header line "Shrunk Sample (<n> steps)"
     - If no shrinking took place samples have header line "Sample"
     - Original samples now also report the original error
-    - Action sequences in stateful properties got their own reporting format
+    - Action sequences in stateful properties got `their own reporting format
 
-- Arbitraries.frequencyOf() now takes covariant arbitrary types
+- `Arbitraries.frequencyOf() now takes covariant arbitrary types
 
 #### Breaking Changes